main: extract run-delivery domain from state.ts
Stage C of the state.ts decomposition: post-run connector delivery
(durable queue, per-connector deliverRunTo* methods, delivery audit)
moves verbatim into run-delivery.ts behind an explicit RunDeliveryHost
interface. AppStateStore keeps one-line delegates, so main.ts, preload,
and tests are untouched. state.ts drops from 6,494 to 5,247 lines.
